backend/ocr_service/requirements.txt
flask==2.3.3
flask-cors==4.0.1
pillow>=10.2.0,<10.3.0
google-generativeai>=0.7.2
protobuf>=4.25.3
requests>=2.32.2
python-dotenv>=1.0.1
